Changed file structure, removed legacy code
