+++ public/index.html
@@ -69,7 +69,7 @@
 
                     <div class = "boxline">
                         Prędkość minimalna [km/h]:
-                        <input type = "number" id = "superelevation_vmin" placeholder = "" >
+                        <input type = "number" id = "superelevation_vmin" placeholder = "80" >
                     </div>
 
                     <div class = "boxline">
@@ -1169,13 +1169,11 @@
             <a href = "https://td2.info.pl/profile/?u=3336">&copy; Ugulele <span id = "year"></span></a>
         </footer>
 
+        <script type="module" src = "js/app.js"></script>
         <script type = "module" src = "js/domutils.js"></script>
-        <script type = "module" src = "js/przechylka.js"></script>
-        <script type = "module" src = "js/curves.js"></script>
-        <script type = "module" src = "js/widening.js"></script>
-        <script type = "module" src = "js/verticalcurves.js"></script>
-        <script type = "module" src = "js/turns.js"></script>
-        <script type = "module" src = "js/functions.js"></script>
+        <!-- <script type = "module" src = "js/curves.js"></script> -->
+        <!-- <script type = "module" src = "js/verticalcurves.js"></script> -->
+        <!-- <script type = "module" src = "js/turns.js"></script> -->
 
     </body>
 </html>
